// frontend/src/generated/api/models/target-update-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Target Management API
 * RESTful API for managing Target objects with full CRUD operations.  ## Features - Full CRUD operations for targets - API versioning (/api/v1/) - Health check endpoint - Request tracing via X-Request-ID header  ## Authentication Currently no authentication required (internal use only). 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface TargetUpdateDTO {
    /**
     * Latitude coordinate
     */
    'latitude'?: number;
    /**
     * Longitude coordinate
     */
    'longitude'?: number;
    /**
     * Altitude in meters
     */
    'altitude'?: number;
    /**
     * Frequency (any positive number)
     */
    'frequency'?: number;
    /**
     * Speed in m/s
     */
    'speed'?: number;
    /**
     * Bearing in degrees
     */
    'bearing'?: number;
    /**
     * IPv4 address
     */
    'ip_address'?: string;
}